requirements.txt

pandas>=2.1
numpy>=1.26
requests>=2.32
python-dateutil>=2.9
PyYAML>=6.0
gspread>=6.0.0
google-auth>=2.33.0
gspread-formatting>=1.2.0
python-dotenv>=1.0.1
markdown>=3.6
